Remove all the cruft lioc_* routines and rename lio_core.c->lio_core_misc.c
